requirements-dev.txt
pyserial==3.5
rope==0.18.0
